swb: WriteBarrierPtr ctor, direct SetVirtualTable
We have initialization `... = {0}`. When the target contains fields of type
WriteBarrierPtr, this calls the default `WriteBarrierPtr()` constructor
for each of such fields. To be compatible the default constructor should
zero itself. Changed default `WriteBarrierPtr()` to zero its wrapped ptr.

However this reveals a side effect of SetVirtualTable implementation. We
use `placement new` to change an object's virtual table. `placement new`
causes constructors of the object and base classes to be called, and in
turn the default constructor of any fields are called. That is not
intended by SetVirtualTable. Above change causes fields to be zeroed
unexpectedly while changing virtual table.

Fixed by changing SetVirtualTable to directly change virtual table value.